move all id classes into single trait
