+++ src/main/scala/ai/metarank/feature/BaseFeature.scala
@@ -3,7 +3,8 @@ package ai.metarank.feature
 import ai.metarank.flow.FieldStore
 import ai.metarank.model.Event.{FeedbackEvent, InteractionEvent, ItemEvent, ItemRelevancy, RankingEvent}
 import ai.metarank.model.FeatureScope.{ItemScope, SessionScope, TenantScope, UserScope}
-import ai.metarank.model.{Event, FeatureSchema, FeatureScope, Field, FieldName, ItemId, MValue, UserId}
+import ai.metarank.model.Identifier._
+import ai.metarank.model.{Event, FeatureSchema, FeatureScope, FieldName, MValue}
 import io.findify.featury.model.Key.{FeatureName, Scope, Tag, Tenant}
 import io.findify.featury.model.{FeatureConfig, FeatureValue, Key, Write}
