+++ src/main/scala/ai/metarank/feature/ItemAgeFeature.scala
@@ -5,7 +5,8 @@ import ai.metarank.feature.ItemAgeFeature.ItemAgeSchema
 import ai.metarank.flow.FieldStore
 import ai.metarank.model.Event.ItemRelevancy
 import ai.metarank.model.FieldName.Item
-import ai.metarank.model.{Event, FeatureSchema, FeatureScope, Field, FieldName, ItemId, MValue, UserId}
+import ai.metarank.model.{Event, FeatureSchema, FeatureScope, Field, FieldName, MValue}
+import ai.metarank.model.Identifier._
 import ai.metarank.model.MValue.SingleValue
 import ai.metarank.util.Logging
 import io.circe.Decoder
