+++ src/main/scala/ai/metarank/feature/UserAgentFeature.scala
@@ -5,7 +5,8 @@ import ai.metarank.feature.UserAgentFeature.UserAgentSchema
 import ai.metarank.feature.ua.{BotField, BrowserField, OSField, PlatformField}
 import ai.metarank.flow.FieldStore
 import ai.metarank.model.Field.{StringField, StringListField}
-import ai.metarank.model.{Event, FeatureSchema, FeatureScope, FieldName, ItemId, MValue, UserId}
+import ai.metarank.model.{Event, FeatureSchema, FeatureScope, FieldName, MValue}
+import ai.metarank.model.Identifier._
 import ai.metarank.model.MValue.VectorValue
 import ai.metarank.util.OneHotEncoder
 import io.circe.{Decoder, DecodingFailure}
