+++ src/test/scala/ai/metarank/feature/UserAgentFeatureTest.scala
@@ -6,8 +6,9 @@ import ai.metarank.feature.ua.PlatformField
 import ai.metarank.model.Event.ItemRelevancy
 import ai.metarank.model.FeatureScope.{ItemScope, SessionScope}
 import ai.metarank.model.Field.StringField
-import ai.metarank.model.{FieldName, ItemId, SessionId}
+import ai.metarank.model.FieldName
 import ai.metarank.model.FieldName.{Item, Ranking}
+import ai.metarank.model.Identifier.SessionId
 import ai.metarank.model.MValue.VectorValue
 import ai.metarank.util.TestRankingEvent
 import cats.data.NonEmptyList
