+++ src/test/scala/ai/metarank/feature/WordCountFeatureTest.scala
@@ -3,10 +3,11 @@ package ai.metarank.feature
 import ai.metarank.feature.WordCountFeature.WordCountSchema
 import ai.metarank.flow.FieldStore
 import ai.metarank.model.Event.ItemRelevancy
-import ai.metarank.model.{FeatureSchema, FieldName, ItemId}
+import ai.metarank.model.{FeatureSchema, FieldName}
 import ai.metarank.model.FeatureScope.ItemScope
 import ai.metarank.model.FieldName.Item
 import ai.metarank.model.Field.StringField
+import ai.metarank.model.Identifier.ItemId
 import ai.metarank.model.MValue.SingleValue
 import ai.metarank.util.{TestMetadataEvent, TestRankingEvent}
 import io.circe.yaml.parser.parse
