+++ src/test/scala/ai/metarank/flow/ImpressionInjectFunctionTest.scala
@@ -1,7 +1,7 @@
 package ai.metarank.flow
 
 import ai.metarank.model.Event.{FeedbackEvent, InteractionEvent, RankingEvent}
-import ai.metarank.model.{EventId, ItemId}
+import ai.metarank.model.EventId
 import ai.metarank.util.{FlinkTest, TestInteractionEvent, TestRankingEvent}
 import org.apache.flink.api.common.RuntimeExecutionMode
 import org.scalatest.flatspec.AnyFlatSpec
