+++ src/test/scala/ai/metarank/util/TestInteractionEvent.scala
@@ -1,7 +1,8 @@
 package ai.metarank.util
 
 import ai.metarank.model.Event.{InteractionEvent, ItemRelevancy}
-import ai.metarank.model.{EventId, Field, ItemId, SessionId, UserId}
+import ai.metarank.model.{EventId, Field}
+import ai.metarank.model.Identifier._
 import io.findify.featury.model.Timestamp
 
 import java.util.UUID
