+++ src/test/scala/ai/metarank/util/TestMetadataEvent.scala
@@ -1,7 +1,8 @@
 package ai.metarank.util
 
 import ai.metarank.model.Event.ItemEvent
-import ai.metarank.model.{EventId, Field, ItemId}
+import ai.metarank.model.Identifier.ItemId
+import ai.metarank.model.{EventId, Field}
 import io.findify.featury.model.Timestamp
 
 import java.util.UUID
