+++ src/test/scala/ai/metarank/util/TestRankingEvent.scala
@@ -1,7 +1,8 @@
 package ai.metarank.util
 
 import ai.metarank.model.Event.{ItemRelevancy, RankingEvent}
-import ai.metarank.model.{EventId, ItemId, SessionId, UserId}
+import ai.metarank.model.EventId
+import ai.metarank.model.Identifier._
 import cats.data.NonEmptyList
 import io.findify.featury.model.Timestamp
